Revert "update"
This reverts commit 5ffa0a1197a37ef375fa73e98b0378272d09d480.

+++ platform/emscripten/Rtt_EmscriptenContext.h
@@ -161,8 +161,8 @@ namespace Rtt
 		std::string fPathToApp;
 		std::string fDocumentsDir;
 
-		float fWidth;
-		float fHeight;
+		int fWidth;
+		int fHeight;
 
 		EmscriptenPlatform* fPlatform;
 		DeviceOrientation::Type fOrientation;
